requirements.txt
pandas
numpy
ensure==1.0.4
python-docx==1.1.2
openpyxl==3.1.3
llama-index-core==0.10.43
langchain
langchain-community
langchain-core
llama-index-readers-web
llama-index
PyPDF2
cohere
pinecone-client
openai